Allow users to overwride Preprocessing.__call__
This allows us to support shared random transformations for inputs and outputs
